Asssignment 7/requirements.txt
Flask==1.1.2
pandas==1.2.4
bs4==0.0.1
lxml==4.6.3
pyOpenSSL==20.0.1
google-cloud-storage==1.37.1